I was able to create a new test for that and i also corrected the defect. The test makes sure that a given statement object doesn't send twice the same query.
I had to mock the MondrianInprocProxy class and create a hash map of the requests that go through it. If the same request is sent twice, an exception gets triggered. The only exception to this is the MDSCHEMA_CUBES query that is used to populate both the catalogs and the schemas associated to a given catalog. This was due to a flaw in SSAS; it doesn't return a SCHEMA_NAME column when asked to. We fixed it this way in some other revision.

git-svn-id: https://olap4j.svn.sourceforge.net/svnroot/olap4j/trunk@127 c6a108a4-781c-0410-a6c6-c2d559e19af0